KAFKA-4148 follow up patch for KIP-79

+++ core/src/main/scala/kafka/server/ReplicaFetcherThread.scala
@@ -264,7 +264,7 @@ class ReplicaFetcherThread(name: String,
   private def earliestOrLatestOffset(topicPartition: TopicPartition, earliestOrLatest: Long, consumerId: Int): Long = {
     val (request, apiVersion) =
       if (brokerConfig.interBrokerProtocolVersion >= KAFKA_0_10_1_IV2) {
-        val partitions = Map(topicPartition -> earliestOrLatest)
+        val partitions = Map(topicPartition -> java.lang.Long.valueOf(earliestOrLatest))
         (new ListOffsetRequest(partitions.asJava, consumerId), 1)
       } else {
         val partitions = Map(topicPartition -> new ListOffsetRequest.PartitionData(earliestOrLatest, 1))
